Change anchor label from 'dandg' to 'drilled'
